requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers & NLP
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.15.0
accelerate>=0.24.0
sentence-transformers>=2.2.0

# Fine-Tuning
peft>=0.6.0
trl>=0.7.0
bitsandbytes>=0.41.0

# LLM Frameworks
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
langgraph>=0.0.20
llama-index>=0.9.0

# Vector Databases
chromadb>=0.4.0
faiss-cpu>=1.7.4
weaviate-client>=3.25.0
pinecone-client>=2.2.0
qdrant-client>=1.7.0

# APIs
openai>=1.3.0
anthropic>=0.7.0
cohere>=4.37

# Agents
crewai>=0.1.0

# Classical NLP
nltk>=3.8.0
spacy>=3.7.0
gensim>=4.3.0
scikit-learn>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.5.0
tqdm>=4.66.0
rich>=13.7.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
ipykernel>=6.27.0

# Evaluation
ragas>=0.1.0
rouge-score>=0.1.2
evaluate>=0.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# APIs & Web
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Development
black>=23.11.0
ruff>=0.1.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Monitoring
wandb>=0.16.0
